requirements.txt
asgiref==3.5.2
bidict==0.22.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==38.0.4
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.5
Flask-SocketIO==5.3.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==2.1.2
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.23
msgpack==1.0.5
numpy==1.24.3
pandas==2.0.1
passlib==1.7.4
protobuf==4.22.3
publicsuffix2==2.20191221
pyasn1==0.5.0
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
requests==2.29.0
ruamel.yaml==0.17.22
scikit-learn==1.2.2
scipy==1.10.1
siphash==0.0.1
six==1.16.0
sortedcontainers==2.4.0
threadpoolctl==3.1.0
tornado==6.3.1
tzdata==2023.3
urllib3==1.26.15
urwid==2.1.2
websocket-client==1.5.1
Werkzeug==2.3.3
wsproto==1.2.0
zstandard==0.19.0
